Backend/requirements.txt
Flask
flask-cors
mysql-connector-python
google-genai
python-dotenv
gunicorn